Add list value support to clrenv 0.2
This is a "feature" that was dropped during the rewrite and the only blocker for switching to it in the primary codebase. Eventually, I would prefer that we transition all the code to str values only to pave the way to a ENV_VAR future, but for now we still have quite a few list-of-primitive values in our environment.